requirements.txt
python-dotenv==0.21.0
uvicorn==0.17.6
fastapi==0.75.2
pydantic==1.9.0
spacy==3.7.5
redis==5.0.8
neo4j==5.28.1
locust==2.15.0